backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.0
supabase==2.0.0
google-generativeai==0.3.1
sentence-transformers==2.2.2
aiohttp==3.8.6
beautifulsoup4==4.12.2
langgraph>=0.0.29 